sync multithreaded refactor
