+++ ferrunix-core/src/object_builder.rs
@@ -1,15 +1,11 @@
 //! Builder singleton or transient objects, with our without dependencies.
 
-#[cfg(all(not(feature = "multithread"), not(feature = "tokio")))]
-#[path = "./object_builder_unsync.rs"]
+#[cfg(not(feature = "tokio"))]
+#[path = "./object_builder_sync.rs"]
 pub(crate) mod inner;
 
 #[cfg(feature = "tokio")]
 #[path = "./object_builder_async.rs"]
-pub(crate) mod inner;
-
-#[cfg(feature = "multithread")]
-#[path = "./object_builder_sync.rs"]
 pub(crate) mod inner;
 
 pub(crate) use inner::*;
@@ -17,8 +13,8 @@ pub(crate) use inner::*;
 /// All possible "objects" that can be held by the registry.
 #[cfg(not(feature = "tokio"))]
 pub(crate) enum Object {
-    Transient(Box<dyn TransientBuilder>),
-    Singleton(Box<dyn SingletonGetter>),
+    Transient(crate::types::BoxedTransientBuilder),
+    Singleton(crate::types::BoxedSingletonGetter),
 }
 
 /// All possible "objects" that can be held by the registry.
